Update to Boot 2.5.2
Boot 2.5.2 required a newer version of Gradle, too, so I updated it
to the current version (7.1 as of this writing).

Also added -initial or -complete to the artifactId field in the
pom.xml files.

Also changed a test, due to a change in the available actuator endpoints
in this version of Boot.